+++ connector/src/test/java/it/eng/connector/integration/negotiation/ContractNegotiationFinalizeIT.java
@@ -165,6 +165,7 @@ public class ContractNegotiationFinalizeIT extends BaseIntegrationTest {
                                 .build()))
                         .build()))
                 .build();
+        agreementRepository.save(agreement);
 
         ContractNegotiation contractNegotiationVerified = ContractNegotiation.Builder.newInstance()
                 .consumerPid(createNewId())